Transfer SPI software implementation to pins.spi
Even though it's only really useful for local implementations it still
belongs there as it's not actually Raspberry Pi specific. Also added
pin_factory args to the underlying SPISoftwareBus class while we're at
it, and fixed up MockFactory to have a new MockSPIInterface based on the
transferred class.
