+++ gpiozero/pins/spi.py
@@ -9,25 +9,126 @@
 import operator
 from threading import RLock
 
+from . import SPI
 from ..devices import Device, SharedMixin
 from ..input_devices import InputDevice
 from ..output_devices import OutputDevice
+from ..exc import DeviceClosed
+
+
+class SPISoftware(SPI):
+    def __init__(self, clock_pin, mosi_pin, miso_pin, select_pin, *,
+                 pin_factory):
+        self._bus = None
+        self._select = None
+        super().__init__(pin_factory=pin_factory)
+        try:
+            # XXX We *should* be storing clock_mode locally, not clock_phase;
+            # after all different users of the bus can disagree about the
+            # clock's polarity and even select pin polarity
+            self._clock_phase = False
+            self._lsb_first = False
+            self._bits_per_word = 8
+            self._bus = SPISoftwareBus(
+                clock_pin, mosi_pin, miso_pin, pin_factory=pin_factory)
+            self._select = OutputDevice(
+                select_pin, active_high=False, pin_factory=pin_factory)
+        except:
+            self.close()
+            raise
+
+    def _conflicts_with(self, other):
+        return not (
+            isinstance(other, SoftwareSPI) and
+            (self._select.pin.number != other._select.pin.number)
+            )
+
+    def close(self):
+        if self._select:
+            self._select.close()
+        self._select = None
+        if self._bus is not None:
+            self._bus.close()
+        self._bus = None
+        super().close()
+
+    @property
+    def closed(self):
+        return self._bus is None
+
+    def __repr__(self):
+        try:
+            self._check_open()
+            return (
+                'SPI(clock_pin={self._bus.clock.pin.number}, '
+                'mosi_pin={self._bus.mosi.pin.number}, '
+                'miso_pin={self._bus.miso.pin.number}, '
+                'select_pin={self._select.pin.number})'.format(self=self))
+        except DeviceClosed:
+            return 'SPI(closed)'
+
+    def transfer(self, data):
+        with self._bus.lock:
+            self._select.on()
+            try:
+                return self._bus.transfer(
+                    data, self._clock_phase, self._lsb_first,
+                    self._bits_per_word)
+            finally:
+                self._select.off()
+
+    def _get_clock_mode(self):
+        with self._bus.lock:
+            return (not self._bus.clock.active_high) << 1 | self._clock_phase
+
+    def _set_clock_mode(self, value):
+        if not (0 <= value < 4):
+            raise SPIInvalidClockMode(
+                "{value} is not a valid clock mode".format(value=value))
+        with self._bus.lock:
+            self._bus.clock.active_high = not (value & 2)
+            self._clock_phase = bool(value & 1)
+
+    def _get_lsb_first(self):
+        return self._lsb_first
+
+    def _set_lsb_first(self, value):
+        self._lsb_first = bool(value)
+
+    def _get_bits_per_word(self):
+        return self._bits_per_word
+
+    def _set_bits_per_word(self, value):
+        if value < 1:
+            raise ValueError('bits_per_word must be positive')
+        self._bits_per_word = int(value)
+
+    def _get_select_high(self):
+        return self._select.active_high
+
+    def _set_select_high(self, value):
+        with self._bus.lock:
+            self._select.active_high = value
+            self._select.off()
 
 
 class SPISoftwareBus(SharedMixin, Device):
-    def __init__(self, clock_pin, mosi_pin, miso_pin):
+    def __init__(self, clock_pin, mosi_pin, miso_pin, *, pin_factory):
         self.lock = None
         self.clock = None
         self.mosi = None
         self.miso = None
         super().__init__()
+        # XXX Should probably just use CompositeDevice for this; would make
+        # close() a bit cleaner - any implications with the RLock?
         self.lock = RLock()
         try:
-            self.clock = OutputDevice(clock_pin, active_high=True)
+            self.clock = OutputDevice(
+                clock_pin, active_high=True, pin_factory=pin_factory)
             if mosi_pin is not None:
-                self.mosi = OutputDevice(mosi_pin)
+                self.mosi = OutputDevice(mosi_pin, pin_factory=pin_factory)
             if miso_pin is not None:
-                self.miso = InputDevice(miso_pin)
+                self.miso = InputDevice(miso_pin, pin_factory=pin_factory)
         except:
             self.close()
             raise
@@ -52,7 +153,7 @@ class SPISoftwareBus(SharedMixin, Device):
         return self.lock is None
 
     @classmethod
-    def _shared_key(cls, clock_pin, mosi_pin, miso_pin):
+    def _shared_key(cls, clock_pin, mosi_pin, miso_pin, *, pin_factory=None):
         return (clock_pin, mosi_pin, miso_pin)
 
     def transfer(self, data, clock_phase=False, lsb_first=False, bits_per_word=8):
